backend/pyproject.toml
[project]
name = "animal-welfare-verifier-api"
version = "0.1.0"
description = "FastAPI backend for balanced reputation search and evidence synthesis."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "crawl4ai>=0.8.6",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pillow>=11.1.0",
  "pydantic-settings>=2.8.0",
  "python-multipart>=0.0.20",
  "python-dotenv>=1.0.1",
  "beautifulsoup4>=4.12.0",
  "facebook-scraper>=0.2.59",
  "trafilatura>=2.0.0",
  "uvicorn[standard]>=0.34.0",
  "apify-client>=2.5.0",
]

[project.optional-dependencies]
llm = [
  "openai>=1.68.0",
]
scrapers = [
  "beautifulsoup4>=4.12.0",
  "facebook-scraper>=0.2.59",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
